v 0.3 *Missing MP check for FastTree *Does not print all tree file names *Now checks if files exist before reading/writing *Removed static reference to the nt reference file *Inconsistant subprocess calls 	-Not following any standard or protocol on which function or how it is run 	-shell = True is used for Windows use, but decreases performance on all platforms
Confirmed to work on both Mac and Windows.

+++ pipeline.py
@@ -23,11 +23,11 @@ else:
 reference = "reference_data/reference.fa"
 
 # ===== Connect your programs here, Linux style
-BLASTN = "/usr/local/bin/ncbi_blast/blastn"
-MAKEBLASTDB = "/usr/local/bin/ncbi_blast/makeblastdb"
+BLASTN = "/Users/michael.zeller/ncbi_blast/blastn"
+MAKEBLASTDB = "/Users/michael.zeller/ncbi_blast/makeblastdb"
 SMOF = "smof"
 MAFFT = "/usr/local/bin/mafft"
-FASTTREE = "/usr/local/bin/FastTree/FastTree"
+FASTTREE = "/Users/michael.zeller/FastTree/FastTree"
 NN_CLASS = "treedist.py"
 PYTHON = "python"
 
@@ -180,13 +180,13 @@ ARR = ["H1", "H3", "N1", "N2", "PB2", "PB1", "PA", "NP", "M", "NS"]
 for segment in ARR:
 	print(segment)
 	segmentFile = outDir + "/" + segment + ".ids"
-	if os.path.isfile(segmentFile):
+	if os.path.isfile(segmentFile)  and os.path.getsize(segmentFile) > 0:
 		#Translators note: import smof, do smoffy things would be better
 		subprocess.run(SMOF + " grep -Xf " + outDir + "/" + segment + ".ids " + baseName + ".clean" + " > " + outDir + "/" + segment + ".fa", shell = True, check = True)
-		subprocess.run(SMOF + " grep \"|" + segment + "|\" reference_data/reference.fa >> query_sample.fasta_output/" + segment + ".fa", shell = True, check = True) 
+		subprocess.run(SMOF + " grep \"|" + segment + "|\" " + reference + " >> " + outDir + "/" + segment + ".fa", shell = True, check = True) 
 		#Cannot get same style to work
 		#subprocess.check_output([SMOF,"grep","\"|" + segment + "|\"", reference, ">>", outDir + "/" + segment + ".fa"], shell = True)
-	os.remove(outDir + "/" + segment + ".ids")
+	#os.remove(outDir + "/" + segment + ".ids")
 	
 # Slow part, building the alignment and tree; slower from shell spin ups
 for segment in ARR:
@@ -197,8 +197,8 @@ for segment in ARR:
 		subprocess.check_output(MAFFT + " --auto --reorder " + outDir + "/" + segment + ".fa > " + outDir + "/" + segment + "_aln.fa", shell = True)
 		#subprocess.check_output([FASTTREE,"-nt","-gtr","-gamma",outDir + "/" + segment + "_aln.fa",">",outDir + "/" + segment + ".tre"], shell = True) # can drop -gtr -gamma for faster results
 		subprocess.check_output(FASTTREE + " -nt -gtr -gamma " + outDir + "/" + segment + "_aln.fa" + " > " + outDir + "/" + segment + ".tre", shell = True) # can drop -gtr -gamma for faster results
-		
-	os.remove(outDir + "/" + segment + ".fa")
+	if(os.path.isfile(outDir + "/" + segment + ".fa")):	
+		os.remove(outDir + "/" + segment + ".fa")
 
 # Fast again, pull out clades
 	finalOutputFile = outDir + "/" + segment + "_Final_Output.txt"
@@ -209,16 +209,26 @@ for segment in ARR:
 # Annotations are based upon reading reference set deflines. For example, H1 genes have
 # the H1 gene at pipe 5, the US HA clade at pipe 1, and the Global HA clade at pipe 8.
 # These positions may be modified, or extended, to return any metadata required.
-subprocess.check_output(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/H1.tre -c 5,1,8 > " + baseName + "_Final_Output.txt", shell = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/H3.tre -c 5,1,8 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/N1.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/N2.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PB2.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PB1.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PA.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/NP.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/M.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
-subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/NS.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/H1.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/H1.tre -c 5,1,8 i>> " + baseName + "_Final_Output.txt", shell = True)
+if os.path.isfile(outDir + "/H3.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/H3.tre -c 5,1,8 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/N1.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/N1.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/N2.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/N2.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/PB2.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PB2.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/PB1.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PB1.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/PA.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/PA.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/NP.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/NP.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/M.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/M.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
+if os.path.isfile(outDir + "/NS.tre"):
+	subprocess.run(PYTHON + " " + NN_CLASS + " " + "-i" + outDir + "/NS.tre -c 5,1 >> " + baseName + "_Final_Output.txt", shell = True, check = True)
 
 copyfile(baseName + "_Final_Output.txt", outDir + "/" + baseName + "_Final_Output.txt")
